requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf_yasg
xhtml2pdf
Jinja2
qrcode